+++ visualization/app/codeCharta/state/store/appSettings/appSettings.reducer.ts
@@ -1,4 +1,5 @@
 // Plop: Append reducer import here
+import { isAttributeSideBarVisible } from "./isAttributeSideBarVisible/isAttributeSideBarVisible.reducer"
 import { sortingOrderAscending } from "./sortingOrderAscending/sortingOrderAscending.reducer"
 import { searchPanelMode } from "./searchPanelMode/searchPanelMode.reducer"
 import { isLoadingFile } from "./isLoadingFile/isLoadingFile.reducer"
@@ -23,6 +24,7 @@ import { combineReducers } from "redux"
 
 const appSettings = combineReducers({
 	// Plop: Append sub-reducer here
+	isAttributeSideBarVisible,
 	sortingOrderAscending,
 	searchPanelMode,
 	isLoadingFile,