+++ visualization/app/codeCharta/ui/attributeTypeSelector/attributeTypeSelector.module.ts
@@ -1,0 +1,7 @@
+import "../../state/state.module"
+import angular from "angular"
+import { attributeTypeSelectorComponent } from "./attributeTypeSelector.component"
+
+angular
+	.module("app.codeCharta.ui.attributeTypeSelector", ["app.codeCharta.state"])
+	.component(attributeTypeSelectorComponent.selector, attributeTypeSelectorComponent)